requirements.txt
africastalking==1.2.8
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
bootstrap4==0.1.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0
django-admin-soft-dashboard==1.0.12
django-bootstrap-datepicker-plus==5.0.5
django-ckeditor-5==0.2.13
django-clearcache==1.2.1
django-jazzmin==3.0.0
django-paypal==2.1
django-taggit==6.1.0
environs==11.0.0
frozenlist==1.4.1
geoip2==4.8.0
google-api-core==2.20.0
google-api-python-client==2.144.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
marshmallow==3.22.0
maxminddb==2.6.2
multidict==6.1.0
packaging==24.1
paypal==1.2.5
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.4
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rsa==4.9
schema==0.7.7
setuptools==75.1.0
shortuuid==1.0.13
sqlparse==0.5.1
stripe==11.0.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.31.0
whitenoise==6.7.0
yarl==1.13.1
